everything moved from gateway.py
