+++ starknet_py/net/schemas/broadcasted_txn.py
@@ -3,8 +3,10 @@ from marshmallow_oneofschema.one_of_schema import OneOfSchema
 
 from starknet_py.net.client_models import TransactionType
 from starknet_py.net.models.transaction import compress_program, decompress_program
-from starknet_py.net.schemas.gateway import SierraCompiledContractSchema
-from starknet_py.net.schemas.rpc.contract import ContractClassSchema
+from starknet_py.net.schemas.rpc.contract import (
+    ContractClassSchema,
+    SierraCompiledContractSchema,
+)
 from starknet_py.net.schemas.rpc.transactions import (
     DeclareTransactionV1Schema,
     DeclareTransactionV2Schema,
